feat(watt): move icon size to an enum
