+++ libs/ui-watt/src/lib/foundations/icon/+storybook/storybook-icon-overview.component.html
@@ -24,19 +24,19 @@ limitations under the License.
 
 <div class="groups custom-icons">
   <figure>
-    <watt-icon name="explore" size="XLarge"></watt-icon>
+    <watt-icon name="explore" [size]="iconSize.XLarge"></watt-icon>
     <figcaption>explore</figcaption>
   </figure>
   <figure>
-    <watt-icon name="power" size="XLarge"></watt-icon>
+    <watt-icon name="power" [size]="iconSize.XLarge"></watt-icon>
     <figcaption>power</figcaption>
   </figure>
   <figure>
-    <watt-icon name="map_marker" size="XLarge"></watt-icon>
+    <watt-icon name="map_marker" [size]="iconSize.XLarge"></watt-icon>
     <figcaption>map_marker</figcaption>
   </figure>
   <figure>
-    <watt-icon name="meter" size="XLarge"></watt-icon>
+    <watt-icon name="meter" [size]="iconSize.XLarge"></watt-icon>
     <figcaption>meter</figcaption>
   </figure>
 </div>
@@ -64,27 +64,27 @@ limitations under the License.
 <h2>Sizes</h2>
 <section class="sizes">
   <figure>
-    <watt-icon name="search" size="XSmall"></watt-icon>
+    <watt-icon name="search" [size]="iconSize.XSmall"></watt-icon>
     <figcaption>XSmall (12px)</figcaption>
   </figure>
   <figure>
-    <watt-icon name="search" size="Small"></watt-icon>
+    <watt-icon name="search" [size]="iconSize.Small"></watt-icon>
     <figcaption>Small (14px)</figcaption>
   </figure>
   <figure>
-    <watt-icon name="search" size="Medium"></watt-icon>
+    <watt-icon name="search" [size]="iconSize.Medium"></watt-icon>
     <figcaption>Medium (16px)</figcaption>
   </figure>
   <figure>
-    <watt-icon name="search" size="Large"></watt-icon>
+    <watt-icon name="search" [size]="iconSize.Large"></watt-icon>
     <figcaption>Large (32px)</figcaption>
   </figure>
   <figure>
-    <watt-icon name="search" size="XLarge"></watt-icon>
+    <watt-icon name="search" [size]="iconSize.XLarge"></watt-icon>
     <figcaption>XLarge (64px)</figcaption>
   </figure>
   <figure>
-    <watt-icon name="search" size="XXLarge"></watt-icon>
+    <watt-icon name="search" [size]="iconSize.XXLarge"></watt-icon>
     <figcaption>XXLarge (128px)</figcaption>
   </figure>
 </section>
