+++ libs/ui-watt/src/lib/foundations/icon/+storybook/storybook-icon-overview.component.ts
@@ -15,7 +15,9 @@
  * limitations under the License.
  */
 import { Component } from '@angular/core';
+
 import { WattIcon } from '../icons';
+import { WattIconSize } from '../watt-icon-size';
 
 interface Icon {
   name: string;
@@ -34,6 +36,11 @@ interface IconGroup {
   styleUrls: ['./storybook-icon-overview.component.scss'],
 })
 export class StorybookIconOverviewComponent {
+  /**
+   * @ignore
+   */
+  iconSize = WattIconSize;
+
   /**
    * @ignore
    */
